Refactor tests into one class * There is only one version of api, no need for base class
